Transfer pybamm model construction methods to BaseModel, Add pybamm SPMe class
